requirements.txt
# Telegram Bot
aiogram>=3.13.0

# OpenAI API
openai>=1.40.0

# Environment & Config
python-dotenv>=1.0.0

# Text-to-Speech
gTTS>=2.5.0

# PDF & Image processing
reportlab>=4.2.0
Pillow>=10.0.0
pdf2image>=1.17.0
pypdf>=4.0.0

# Async & Logging
aiofiles>=24.1.0
